feat: add global exception handling middleware for clean JSON error responses
Implemented ExceptionHandlingMiddleware to standardize API error responses in JSON format
Added specific handling for InvalidRoleAssignmentException with HTTP 403 response
Updated Program.cs to register and use custom middleware
Improved logging for both handled and unhandled exceptions

+++ Stamp.Web/Middleware/ExceptionHandlingMiddleware.cs
@@ -43,7 +43,7 @@ namespace Stamp.Web.Middleware
 
             var errorResponse = new
             {
-                status = (int)statusCode,
+                status = statusCode,
                 error = message,
                 traceId = context.TraceIdentifier
             };
